[zhenhui] #68 map to local reason

+++ app/src/main/java/org/openlmis/core/network/model/StockCardsNetworkResponse.java
@@ -42,7 +42,7 @@ public class StockCardsNetworkResponse {
   public static class StockMovementItemResponse {
 
     private int movementQuantity;
-    private int requested;
+    private Long requested;
     private long processedDate;
     private String type;
     private String stockOnHand;
